+++ programs/lb_clmm/src/utils/pda.rs
@@ -1,5 +1,6 @@
 use super::seeds::{self, BIN_ARRAY, BIN_ARRAY_BITMAP_SEED, ORACLE, PRESET_PARAMETER};
 use anchor_lang::prelude::Pubkey;
+use num_traits::ToBytes;
 use std::{cmp::max, cmp::min};
 
 pub fn derive_lb_pair_pda2(
@@ -100,6 +101,18 @@ pub fn derive_event_authority_pda() -> (Pubkey, u8) {
     Pubkey::find_program_address(&[b"__event_authority"], &crate::ID)
 }
 
+#[deprecated]
 pub fn derive_preset_parameter_pda(bin_step: u16) -> (Pubkey, u8) {
     Pubkey::find_program_address(&[PRESET_PARAMETER, &bin_step.to_le_bytes()], &crate::ID)
+}
+
+pub fn derive_preset_parameter_pda2(bin_step: u16, base_factor: u16) -> (Pubkey, u8) {
+    Pubkey::find_program_address(
+        &[
+            PRESET_PARAMETER,
+            &bin_step.to_le_bytes(),
+            &base_factor.to_le_bytes(),
+        ],
+        &crate::ID,
+    )
 }